change properties double to float

+++ Models/DadosLancamento.cs
@@ -12,17 +12,17 @@ namespace calculadora_api.Models
 
         public string contractRef { get; set; }
         public string formDataCalculo { get; set; }
-        public double formMulta { get; set; }
-        public double formJuros { get; set; }
-        public double formHonorarios { get; set; }
-        public double formMultaSobContrato { get; set; }
+        public float formMulta { get; set; }
+        public float formJuros { get; set; }
+        public float formHonorarios { get; set; }
+        public float formMultaSobContrato { get; set; }
         public string formIndice { get; set; }
-        public double formIndiceEncargos { get; set; }
+        public float formIndiceEncargos { get; set; }
         public string dataVencimento { get; set; }
-        public double saldoDevedor { get; set; }
+        public float saldoDevedor { get; set; }
         public string tipoLancamento { get; set; }
         public string dataBaseAtual { get; set; }
-        public double valorLancamento { get; set; }
+        public float valorLancamento { get; set; }
